refactor: enum dispatch
