+++ Cargo.toml
@@ -20,6 +20,7 @@ bitflags = "2.0.2"
 multiboot2 = "0.14.0"
 rustc-demangle = "0.1.21"
 acpi = "4.1.1"
+enum_dispatch = "0.3.12"
 time = { version = "0.3.28", default-features = false }
 kfs_macro = { path = "./kfs_macro" }
