+++ src/fs/syscall/chown.rs
@@ -1,3 +1,4 @@
+use crate::fs::vfs::RealEntry;
 use crate::fs::{path::Path, vfs::lookup_entry_follow};
 use crate::mm::user::verify::verify_path;
 use crate::{process::task::CURRENT, syscall::errno::Errno};
