+++ src/net/local/stream.rs
@@ -185,7 +185,7 @@ impl Socket for LocalStreamSocket {
 		let ent = lookup_entry_follow(&addr.path, task)?;
 
 		let sock = match ent {
-			VfsRealEntry::Socket(sock) => sock,
+			VfsRealEntry::ArcVfsSocketEntry(sock) => sock,
 			_ => return Err(Errno::ECONNREFUSED),
 		};
